🌱 nit: split network-types to improve maintenance
This change split the inflated network-types.go into smaller go files
with each component from NetworkSpec, targeting to improve the
maintenance while increasing the project features.

The motivation for this change is that recently many new capabilities have
been added to network, VPC, subnets, load balancers, and the file
increased by size, placing the type definition, and its methods, from
many network components. Splitting into smaller files also would provide
the project navigation (and understanding for new contributors), and
increase test coverage.
