+++ api/v1beta2/awscluster_types.go
@@ -166,113 +166,6 @@ type Bastion struct {
 	AMI string `json:"ami,omitempty"`
 }
 
-// LoadBalancerType defines the type of load balancer to use.
-type LoadBalancerType string
-
-var (
-	// LoadBalancerTypeClassic is the classic ELB type.
-	LoadBalancerTypeClassic = LoadBalancerType("classic")
-	// LoadBalancerTypeELB is the ELB type.
-	LoadBalancerTypeELB = LoadBalancerType("elb")
-	// LoadBalancerTypeALB is the ALB type.
-	LoadBalancerTypeALB = LoadBalancerType("alb")
-	// LoadBalancerTypeNLB is the NLB type.
-	LoadBalancerTypeNLB = LoadBalancerType("nlb")
-	// LoadBalancerTypeDisabled disables the load balancer.
-	LoadBalancerTypeDisabled = LoadBalancerType("disabled")
-)
-
-// AWSLoadBalancerSpec defines the desired state of an AWS load balancer.
-type AWSLoadBalancerSpec struct {
-	// Name sets the name of the classic ELB load balancer. As per AWS, the name must be unique
-	// within your set of load balancers for the region, must have a maximum of 32 characters, must
-	// contain only alphanumeric characters or hyphens, and cannot begin or end with a hyphen. Once
-	// set, the value cannot be changed.
-	// +kubebuilder:validation:MaxLength:=32
-	// +kubebuilder:validation:Pattern=`^[A-Za-z0-9]([A-Za-z0-9]{0,31}|[-A-Za-z0-9]{0,30}[A-Za-z0-9])$`
-	// +optional
-	Name *string `json:"name,omitempty"`
-
-	// Scheme sets the scheme of the load balancer (defaults to internet-facing)
-	// +kubebuilder:default=internet-facing
-	// +kubebuilder:validation:Enum=internet-facing;internal
-	// +optional
-	Scheme *ELBScheme `json:"scheme,omitempty"`
-
-	// CrossZoneLoadBalancing enables the classic ELB cross availability zone balancing.
-	//
-	// With cross-zone load balancing, each load balancer node for your Classic Load Balancer
-	// distributes requests evenly across the registered instances in all enabled Availability Zones.
-	// If cross-zone load balancing is disabled, each load balancer node distributes requests evenly across
-	// the registered instances in its Availability Zone only.
-	//
-	// Defaults to false.
-	// +optional
-	CrossZoneLoadBalancing bool `json:"crossZoneLoadBalancing"`
-
-	// Subnets sets the subnets that should be applied to the control plane load balancer (defaults to discovered subnets for managed VPCs or an empty set for unmanaged VPCs)
-	// +optional
-	Subnets []string `json:"subnets,omitempty"`
-
-	// HealthCheckProtocol sets the protocol type for ELB health check target
-	// default value is ELBProtocolSSL
-	// +kubebuilder:validation:Enum=TCP;SSL;HTTP;HTTPS;TLS;UDP
-	// +optional
-	HealthCheckProtocol *ELBProtocol `json:"healthCheckProtocol,omitempty"`
-
-	// HealthCheck sets custom health check configuration to the API target group.
-	// +optional
-	HealthCheck *TargetGroupHealthCheckAPISpec `json:"healthCheck,omitempty"`
-
-	// AdditionalSecurityGroups sets the security groups used by the load balancer. Expected to be security group IDs
-	// This is optional - if not provided new security groups will be created for the load balancer
-	// +optional
-	AdditionalSecurityGroups []string `json:"additionalSecurityGroups,omitempty"`
-
-	// AdditionalListeners sets the additional listeners for the control plane load balancer.
-	// This is only applicable to Network Load Balancer (NLB) types for the time being.
-	// +listType=map
-	// +listMapKey=port
-	// +optional
-	AdditionalListeners []AdditionalListenerSpec `json:"additionalListeners,omitempty"`
-
-	// IngressRules sets the ingress rules for the control plane load balancer.
-	// +optional
-	IngressRules []IngressRule `json:"ingressRules,omitempty"`
-
-	// LoadBalancerType sets the type for a load balancer. The default type is classic.
-	// +kubebuilder:default=classic
-	// +kubebuilder:validation:Enum:=classic;elb;alb;nlb;disabled
-	LoadBalancerType LoadBalancerType `json:"loadBalancerType,omitempty"`
-
-	// DisableHostsRewrite disabled the hair pinning issue solution that adds the NLB's address as 127.0.0.1 to the hosts
-	// file of each instance. This is by default, false.
-	DisableHostsRewrite bool `json:"disableHostsRewrite,omitempty"`
-
-	// PreserveClientIP lets the user control if preservation of client ips must be retained or not.
-	// If this is enabled 6443 will be opened to 0.0.0.0/0.
-	PreserveClientIP bool `json:"preserveClientIP,omitempty"`
-}
-
-// AdditionalListenerSpec defines the desired state of an
-// additional listener on an AWS load balancer.
-type AdditionalListenerSpec struct {
-	// Port sets the port for the additional listener.
-	// +kubebuilder:validation:Minimum=1
-	// +kubebuilder:validation:Maximum=65535
-	Port int64 `json:"port"`
-
-	// Protocol sets the protocol for the additional listener.
-	// Currently only TCP is supported.
-	// +kubebuilder:validation:Enum=TCP
-	// +kubebuilder:default=TCP
-	Protocol ELBProtocol `json:"protocol,omitempty"`
-
-	// HealthCheck sets the optional custom health check configuration to the API target group.
-	// +optional
-	HealthCheck *TargetGroupHealthCheckAdditionalSpec `json:"healthCheck,omitempty"`
-}
-
 // AWSClusterStatus defines the observed state of AWSCluster.
 type AWSClusterStatus struct {
 	// +kubebuilder:default=false
