+++ api/v1beta2/network_securitygroup_types.go
@@ -1,0 +1,214 @@
+/*
+Copyright 2022 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1beta2
+
+import (
+	"fmt"
+	"sort"
+)
+
+// SecurityGroupRole defines the unique role of a security group.
+// +kubebuilder:validation:Enum=bastion;node;controlplane;apiserver-lb;lb;node-eks-additional
+type SecurityGroupRole string
+
+var (
+	// SecurityGroupBastion defines an SSH bastion role.
+	SecurityGroupBastion = SecurityGroupRole("bastion")
+
+	// SecurityGroupNode defines a Kubernetes workload node role.
+	SecurityGroupNode = SecurityGroupRole("node")
+
+	// SecurityGroupEKSNodeAdditional defines an extra node group from eks nodes.
+	SecurityGroupEKSNodeAdditional = SecurityGroupRole("node-eks-additional")
+
+	// SecurityGroupControlPlane defines a Kubernetes control plane node role.
+	SecurityGroupControlPlane = SecurityGroupRole("controlplane")
+
+	// SecurityGroupAPIServerLB defines a Kubernetes API Server Load Balancer role.
+	SecurityGroupAPIServerLB = SecurityGroupRole("apiserver-lb")
+
+	// SecurityGroupLB defines a container for the cloud provider to inject its load balancer ingress rules.
+	SecurityGroupLB = SecurityGroupRole("lb")
+)
+
+// SecurityGroup defines an AWS security group.
+type SecurityGroup struct {
+	// ID is a unique identifier.
+	ID string `json:"id"`
+
+	// Name is the security group name.
+	Name string `json:"name"`
+
+	// IngressRules is the inbound rules associated with the security group.
+	// +optional
+	IngressRules IngressRules `json:"ingressRule,omitempty"`
+
+	// Tags is a map of tags associated with the security group.
+	Tags Tags `json:"tags,omitempty"`
+}
+
+// String returns a string representation of the security group.
+func (s *SecurityGroup) String() string {
+	return fmt.Sprintf("id=%s/name=%s", s.ID, s.Name)
+}
+
+// SecurityGroupProtocol defines the protocol type for a security group rule.
+type SecurityGroupProtocol string
+
+var (
+	// SecurityGroupProtocolAll is a wildcard for all IP protocols.
+	SecurityGroupProtocolAll = SecurityGroupProtocol("-1")
+
+	// SecurityGroupProtocolIPinIP represents the IP in IP protocol in ingress rules.
+	SecurityGroupProtocolIPinIP = SecurityGroupProtocol("4")
+
+	// SecurityGroupProtocolTCP represents the TCP protocol in ingress rules.
+	SecurityGroupProtocolTCP = SecurityGroupProtocol("tcp")
+
+	// SecurityGroupProtocolUDP represents the UDP protocol in ingress rules.
+	SecurityGroupProtocolUDP = SecurityGroupProtocol("udp")
+
+	// SecurityGroupProtocolICMP represents the ICMP protocol in ingress rules.
+	SecurityGroupProtocolICMP = SecurityGroupProtocol("icmp")
+
+	// SecurityGroupProtocolICMPv6 represents the ICMPv6 protocol in ingress rules.
+	SecurityGroupProtocolICMPv6 = SecurityGroupProtocol("58")
+
+	// SecurityGroupProtocolESP represents the ESP protocol in ingress rules.
+	SecurityGroupProtocolESP = SecurityGroupProtocol("50")
+)
+
+// IngressRule defines an AWS ingress rule for security groups.
+type IngressRule struct {
+	// Description provides extended information about the ingress rule.
+	Description string `json:"description"`
+	// Protocol is the protocol for the ingress rule. Accepted values are "-1" (all), "4" (IP in IP),"tcp", "udp", "icmp", and "58" (ICMPv6), "50" (ESP).
+	// +kubebuilder:validation:Enum="-1";"4";tcp;udp;icmp;"58";"50"
+	Protocol SecurityGroupProtocol `json:"protocol"`
+	// FromPort is the start of port range.
+	FromPort int64 `json:"fromPort"`
+	// ToPort is the end of port range.
+	ToPort int64 `json:"toPort"`
+
+	// List of CIDR blocks to allow access from. Cannot be specified with SourceSecurityGroupID.
+	// +optional
+	CidrBlocks []string `json:"cidrBlocks,omitempty"`
+
+	// List of IPv6 CIDR blocks to allow access from. Cannot be specified with SourceSecurityGroupID.
+	// +optional
+	IPv6CidrBlocks []string `json:"ipv6CidrBlocks,omitempty"`
+
+	// The security group id to allow access from. Cannot be specified with CidrBlocks.
+	// +optional
+	SourceSecurityGroupIDs []string `json:"sourceSecurityGroupIds,omitempty"`
+
+	// The security group role to allow access from. Cannot be specified with CidrBlocks.
+	// The field will be combined with source security group IDs if specified.
+	// +optional
+	SourceSecurityGroupRoles []SecurityGroupRole `json:"sourceSecurityGroupRoles,omitempty"`
+}
+
+// String returns a string representation of the ingress rule.
+func (i IngressRule) String() string {
+	return fmt.Sprintf("protocol=%s/range=[%d-%d]/description=%s", i.Protocol, i.FromPort, i.ToPort, i.Description)
+}
+
+// IngressRules is a slice of AWS ingress rules for security groups.
+type IngressRules []IngressRule
+
+// Difference returns the difference between this slice and the other slice.
+func (i IngressRules) Difference(o IngressRules) (out IngressRules) {
+	for index := range i {
+		x := i[index]
+		found := false
+		for oIndex := range o {
+			y := o[oIndex]
+			if x.Equals(&y) {
+				found = true
+				break
+			}
+		}
+
+		if !found {
+			out = append(out, x)
+		}
+	}
+
+	return
+}
+
+// Equals returns true if two IngressRule are equal.
+func (i *IngressRule) Equals(o *IngressRule) bool {
+	// ipv4
+	if len(i.CidrBlocks) != len(o.CidrBlocks) {
+		return false
+	}
+
+	sort.Strings(i.CidrBlocks)
+	sort.Strings(o.CidrBlocks)
+
+	for i, v := range i.CidrBlocks {
+		if v != o.CidrBlocks[i] {
+			return false
+		}
+	}
+	// ipv6
+	if len(i.IPv6CidrBlocks) != len(o.IPv6CidrBlocks) {
+		return false
+	}
+
+	sort.Strings(i.IPv6CidrBlocks)
+	sort.Strings(o.IPv6CidrBlocks)
+
+	for i, v := range i.IPv6CidrBlocks {
+		if v != o.IPv6CidrBlocks[i] {
+			return false
+		}
+	}
+
+	if len(i.SourceSecurityGroupIDs) != len(o.SourceSecurityGroupIDs) {
+		return false
+	}
+
+	sort.Strings(i.SourceSecurityGroupIDs)
+	sort.Strings(o.SourceSecurityGroupIDs)
+
+	for i, v := range i.SourceSecurityGroupIDs {
+		if v != o.SourceSecurityGroupIDs[i] {
+			return false
+		}
+	}
+
+	if i.Description != o.Description || i.Protocol != o.Protocol {
+		return false
+	}
+
+	// AWS seems to ignore the From/To port when set on protocols where it doesn't apply, but
+	// we avoid serializing it out for clarity's sake.
+	// See: https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_IpPermission.html
+	switch i.Protocol {
+	case SecurityGroupProtocolTCP,
+		SecurityGroupProtocolUDP,
+		SecurityGroupProtocolICMP,
+		SecurityGroupProtocolICMPv6:
+		return i.FromPort == o.FromPort && i.ToPort == o.ToPort
+	case SecurityGroupProtocolAll, SecurityGroupProtocolIPinIP, SecurityGroupProtocolESP:
+		// FromPort / ToPort are not applicable
+	}
+
+	return true
+}
